+++ src/main/java/com/alex/technicaltest/domain/port/PriceRepository.java
@@ -6,8 +6,8 @@ import com.alex.technicaltest.domain.model.Price;
 
 public interface PriceRepository {
     List<Price> getAll();
-    Price getById(Integer id);
+    Price getById(Long id);
     Price create(Price price);
     Price update(Price price);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
